Implement snapshots in Filesys::POSIX::Snapshot
Case 53875: This commit adds a variety of code cleanups, refactorings, and fresh
code that came about in the course of implementing a snapshot filesystem type.

Filesys::POSIX::Snapshot provides a mechanism whereby a user can mount a
snapshot of an arbitrary point in the filesystem to any location within the
filesystem; one can create a snapshot of /, or even a previously-made snapshot,
with this mechanism, making it possible to create "scratch pad" type
environments for performing work to create temporary and incremental filesystem
modifications.

This mechanism was primarily created to aid in the writing of code that performs
temporary modifications to files brought in from the real filesystem by use of
Filesys::POSIX::Real, without actually modifying any real underlying files.
Previously, complex techniques such as using Filesys::POSIX::Extensions->map()
to precariously map files and directories in a piecemeal manner into scratch pad
areas in such a way that modifications are not made to the real files, were
necessary.  However, now one can perform temporary modifications of real files
and entire directory structures in a fairly native way, by simply mounting a
snapshot of any part of a filesystem, and performing modifications upon the data
in the snapshot as needed.