Correct handling files with names that evaluate false.
Case 75761: Filesys::POSIX had a few places in the code that
mishandled filenames like "0", treating them as if they
represented the end of the file list.

In some cases, this could cause incomplete archives to be
created.

Correct this problem by using definedness instead of truth
checks. Unit tests have been added which verify the fix.