Avoid blocking when taring a FS with named pipes
Case 82969: Added the $O_NONBLOCK flag to prevent named pipes (FIFOs) from
blocking when they are opened to create a tarball.

Unfortunately, the system value for O_NONBLOCK can (and does) vary from
the value in Filesys::POSIX::Bits, so several conversion functions were
written to convert the Bits value to the system value for the constants.
Any functions that interact with real files, etc. on the system now use
these conversions.

Finally, the constant $O_SYMLINK, which was exported but never defined,
has been removed.