Don't create archives with mismatching file sizes.
Case 90017: Because of the way tar files are constructed, and
the way Filesys::POSIX buffers its data, there is a race
condition that makes it possible to have written a header for a
file to disk and later find out that the file size listed in
the header is no longer correct, either because someone has
appended to the file or because someone has truncated the file.

Previously, the Filesys::POSIX::Userland::Tar code did a poor job
of handling this condition (because it was unaware of it) and
ended up producing corrupt archives which could not be extracted.

This commit addresses both the append and the truncate problem,
but in two different ways. For example:

    - If a file's header was written to indicate that the file was
      N bytes long, and it is later found to be N + 1,000 bytes
      long, only the first N bytes are used, even though more were
      available. (The previous behavior was to silently produce a
      corrupt archive.)

    - If a file's header was written to indicate that the file was
      N bytes long, and it is later found to be N - 1,000 bytes
      long, Filesys::POSIX will pad the remaining portion of the
      file with null bytes and produce a warning. (The previous
      behavior was to silently produce a corrupt archive.)

This change is analogous to one that will most likely also be made
in Archive::Tar::Builder in a subcase.