Improve EOF detection when reading files for inclusion in tar.
Case 92573: The EOF detection added in case 90017 is not as
robust as it should have been. This commit enhances the read
loop to ensure that one of the following conditions is always
met:
  1. We have a full buffer to write to a block.
  2. We have a buffer that matches the exact number of predicted
     bytes remaining in the file.
  3. We have unexpectedly reached the EOF (signified by read()
     returning zero).

The undesirable 4th condition that could previously be reached
was that we had a partial buffer even though read() returned
a value greater than zero.

Unit tests have also been added for the three basic scenarios
of concern:
  1. A normal file being archived is preserved as expected.
  2. A file that shrinks relative to the size reported in the
     header is successfully padded with null bytes to match
     the size the header reported.
  3. A file that grows relative to the size reported in the
     header is only read/archived up to the originally reported
     length.